Conflicts:
	Eduras/src/de/illonis/eduras/events/GameEvent.java
	Eduras/src/de/illonis/eduras/logic/ServerEventTriggerer.java
	Eduras/src/de/illonis/eduras/logicabstraction/EdurasInitializer.java

+++ Eduras/src/de/illonis/eduras/items/weapons/SniperWeapon.java
@@ -27,9 +27,8 @@ public class SniperWeapon extends Weapon {
 	public SniperWeapon(GameInformation gi, TimingSource timingSource, int id) {
 		super(ObjectType.ITEM_WEAPON_SNIPER, gi, timingSource, id);
 		setName("Sniper");
-		setShape(new Circle((float) S.go_sniperweapon_shape_radius,
-				(float) S.go_sniperweapon_shape_radius,
-				(float) S.go_sniperweapon_shape_radius));
+		setShape(new Circle(S.go_sniperweapon_shape_radius,
+				S.go_sniperweapon_shape_radius, S.go_sniperweapon_shape_radius));
 		defaultCooldown = S.go_sniperweapon_cooldown;
 		setAmmunitionLimited(S.go_sniperweapon_fillamount,
 				S.go_sniperweapon_maxammo);
